requirements.txt
bing-image-downloader
opencv-python-headless
Pillow
tqdm
torch
torchvision
requests
streamlit
FPDF
